// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        DistrictAdmin: {
            readonly id: number;
            name: string;
            region?: string;
        };
// ...
        ParishAdmin: {
            readonly id: number;
            subcounty: number;
            readonly subcounty_name: string;
            readonly district_name: string;
            name: string;
            agro_zone?: string;
            /** Format: int64 */
            lot_min_bags?: number;
        };
// ...
        SubcountyAdmin: {
            readonly id: number;
            district: number;
            readonly district_name: string;
            name: string;
        };
// ...
        SystemUserAdmin: {
            readonly id: number;
            phone: string;
            full_name: string;
            role: components["schemas"]["RoleEnum"];
            scope_level: components["schemas"]["SystemUserAdminScopeLevelEnum"];
            /** Format: int64 */
            scope_id?: number | null;
            is_active?: boolean;
            password?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}